--- a/getgauge/static_loader.py
+++ b/getgauge/static_loader.py
@@ -1,11 +1,10 @@
 import logging
 import re
-from os import path
+import os
 
 from redbaron import RedBaron
 
 from getgauge.registry import registry
-from getgauge.util import *
 
 
 def load_steps(ast, file_name):
@@ -35,8 +34,9 @@ def reload_steps(content, file_name):
 
 def _create_span(func):
     try:
-        start = func.absolute_bounding_box.top_left
-        end = func.absolute_bounding_box.bottom_right
+        absolute_bounding_box = func.absolute_bounding_box
+        start = absolute_bounding_box.top_left
+        end = absolute_bounding_box.bottom_right
         return {"start": start.line, "startChar": start.column, "end": end.line, "endChar": end.column}
     except:
         return {"start": 0, "startChar": 0, "end": 0, "endChar": 0}
@@ -58,5 +58,5 @@ def load_files(step_impl_dir):
             if ast is not None:
                 load_steps(ast, file_path)
             impl_file.close()
-        elif path.isdir(file_path):
+        elif os.path.isdir(file_path):
             load_files(file_path)