optimising initialisation

Execution time for the sample attached in #54 
Before:
gauge run specs  125.13s user 0.92s system 97% cpu 2:09.64 total
After:
gauge run specs  5.18s user 0.82s system 87% cpu 6.879 total