@BugDiver 2018-05-24
The `create_span` method tries to access the `boundind_box` of the node, which is calculated lazily. Since we were calculating the location for all the steps upfront(while loading the files) it was taking a lot of time. Now the location is calculated only when needed (for stepPositionRequest).